etl_lambda/requirements.txt
pandas
psycopg2-binary
yfinance
boto3
numpy<2.0.0